requirements.txt
chromedriver_autoinstaller==0.6.2
openpyxl==3.1.2
pandas==1.4.4
Pillow==9.5.0
Requests==2.31.0
selenium==4.14.0
sql_metadata==2.9.0
sqlparse==0.4.4
urllib3==2.0.6
